requirements.txt
amqp==2.2.2
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.5.0.3
celery==4.1.0
celery-with-redis==3.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.1
django-celery==3.2.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-pure-pagination==0.3.0
django-redis==4.9.0
django-reversion==2.0.13
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.7
jdcal==1.4
kombu==4.1.0
mysqlclient==1.3.13
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.2.0
pycparser==2.18
pycryptodomex==3.6.4
PyMySQL==0.9.3
python-alipay-sdk==1.8.0
pytz==2018.9
PyYAML==3.13
redis==2.10.6
requests==2.21.0
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
vine==1.2.0
xlrd==1.2.0
xlwt==1.3.0
